test: extract provider/tracker validation suite from AppStateTests (#431)
Continues the AppStateTests decomposition (slice 1 was #452, the issue-
export cluster). Moves the 14 provider/tracker validation tests — API-key
validation, GitHub repository validation/loading, and Jira project/issue-
type validation and refresh — verbatim into a focused
AppStateProviderValidationTests suite.

Pure test reorganization: assertions relocated byte-for-byte, no
production code touched, AppStateHarness + mocks shared via TestSupport.
AppStateTests drops 77 -> 63 methods; the new suite holds 14; total
unchanged (zero coverage loss). Full BugNarratorTests target green.

On inspection the other planned clusters proved thin against current main:
the session-library copy/save/delete behavior is already covered directly
by SessionLibraryControllerTests (only one delete test routes through
AppState), and the permission tests are coupled to startSession — i.e. the
recording-flow orchestration that the plan wants to KEEP as the AppState
smoke suite. So the provider-validation cluster was the substantive
remaining extraction; AppStateTests is now the intended SwiftUI-facing
orchestration smoke suite.

Closes #431
